*New Documents : -IMOS_user_code_library_DRAFT_2012_02_18_V1_for_finalReview Finalise version of the first Version of the document. Needs final review and approval amongts people invoveld in the task
-IMOS_user_code_library_DRAFT_2012_02_18_V1_with_exportCSV_for_review
new tool which uses the netCDFParse. This function described in this document is suppose to export a parsed NetCDF into a CSV file. Needs review of the document

*Scripts
NetCDFParse : update some minor bugs
ouputCSV : new function to export a parsed netcdf to a CSV. Needs to be tested under Linux and Windows. Needs also feedbacks to improve it